requirements.txt
fastapi
uvicorn
pika
grpcio
grpcio-tools
aiofiles
python-multipart